(svn r8973) -Feature/Codechange: Changed the internal unit for aircraft velocities to from 8mph to 1km/h (peter1138), also give aircraft realsitic velocities (so that 1km/h is 1km/h independent of vehicle type) (peter1138). Introduce a flag to reduce aircraft speed in holding patterns to some realistic velocity. Use this flag for the city airport for the time being and make use of the different entry points for this airport type.

+++ src/engine.h
@@ -68,7 +68,7 @@ typedef struct AircraftVehicleInfo {
 	byte subtype;
 	SoundFxByte sfx;
 	byte acceleration;
-	byte max_speed;
+	uint16 max_speed;
 	byte mail_capacity;
 	uint16 passenger_capacity;
 } AircraftVehicleInfo;
